Progroess Bar for each Iteratator
